aarch64: Add test of rusty-loader
This test compiles the rusty-loader with an application artifact compiled from
https://github.com/jschwe/libhermit-rs/commit/f5faa6d059f29e6a4d28a8872acf639c64d3a6ae
The artifact is technically the rusty_demo artifact, since the hello_world artifact
didn't work for me.
This is most likely due to the loader still being brittle.
I suspect that the copying process is buggy, since (un)commenting simple things
such as log messages in rusty-loader will break or fix entry into the kernel.
Until this is fixed this workflow will probably fail sometimes.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>